+++ src/gemseo_fmu/disciplines/time_stepping_system.py
@@ -21,6 +21,7 @@ from typing import TYPE_CHECKING
 from typing import Any
 
 from gemseo.core._process_flow.base_process_flow import BaseProcessFlow
+from gemseo.core.coupling_structure import CouplingStructure
 from gemseo.core.discipline.discipline import Discipline
 from gemseo.mda.mda_chain import MDAChain
 from gemseo.utils.constants import READ_ONLY_EMPTY_DICT
@@ -81,14 +82,17 @@ class TimeSteppingSystem(Discipline):
     __fmu_discipline: list[FMUDiscipline]
     """The FMU disciplines."""
 
+    __mda: MDAChain
+    """The MDA defining the master algorithm to co-simulate the discipline."""
+
+    __mda_max_iter_at_t0: int
+    """ The maximum number of iterations of the MDA algorithm at initial time."""
+
     __restart: bool
     """Whether the system starts from the initial time at each execution."""
 
     __time_manager: TimeManager
     """The time manager."""
-
-    __mda: MDAChain
-    """The MDA defining the master algorithm to co-simulate the discipline."""
 
     _process_flow_class = _TimeSteppingSystemProcessFlow
 
@@ -102,6 +106,7 @@ class TimeSteppingSystem(Discipline):
         do_step: bool = False,
         mda_name: str = "MDAJacobi",
         mda_options: Mapping[str, Any] = READ_ONLY_EMPTY_DICT,
+        mda_max_iter_at_t0: int = 0,
         **fmu_options: Any,
     ) -> None:
         """
@@ -121,11 +126,14 @@ class TimeSteppingSystem(Discipline):
                 Otherwise, simulate the model from initial time to `final_time`.
             mda_name: The MDA class name.
             mda_options: The options of the MDA.
+            mda_max_iter_at_t0: The maximum number of iterations of the MDA algorithm
+                at initial time, to find a multidisciplinary feasible configuration.
             **fmu_options: The options to instantiate the FMU disciplines.
         """  # noqa: D205 D212 D415
         self.__do_step = do_step
         self.__time_manager = TimeManager(0.0, final_time, time_step)
         self.__restart = restart
+        self.__mda_max_iter_at_t0 = mda_max_iter_at_t0
         discipline_time_step = time_step if apply_time_step_to_disciplines else 0.0
         all_disciplines = []
         for discipline in disciplines:
@@ -151,6 +159,16 @@ class TimeSteppingSystem(Discipline):
             if isinstance(discipline, BaseFMUDiscipline)
         ]
         super().__init__()
+        strong_couplings = CouplingStructure(all_disciplines).strong_couplings
+        for fmu_discipline in self.__fmu_disciplines:
+            input_grammar = fmu_discipline.input_grammar
+            array_strong_couplings = set()
+            for strong_coupling in strong_couplings:
+                if strong_coupling in input_grammar:
+                    array_strong_couplings.add(strong_coupling)
+
+            input_grammar.update_from_names(array_strong_couplings)
+
         self.__mda = MDAChain(
             all_disciplines,
             inner_mda_name=mda_name,
@@ -158,6 +176,22 @@ class TimeSteppingSystem(Discipline):
             max_mda_iter=0,
             **mda_options,
         )
+        for mda in self.__mda.inner_mdas:
+            mda.settings.max_mda_iter = 0
+
+        # TimeSteppingSystem uses an MDA
+        # to co-simulate from t(k) to t(k+1) and
+        # to iterate at t(0) to find a multidisciplinary starting state.
+        # If the coupled system is an n-order ODE, with n > 1,
+        # the coupling vector can remain constant during one iteration
+        # before changing at the next iteration.
+        # For this reason, we need to disable the caches.
+        cache_type = self.__mda.CacheType.NONE
+        self.__mda.set_cache(cache_type)
+        self.__mda.mdo_chain.set_cache(cache_type)
+        for mda in self.__mda.inner_mdas:
+            mda.set_cache(cache_type)
+
         self.input_grammar.update(self.__mda.input_grammar)
         self.output_grammar.update(self.__mda.output_grammar)
 
@@ -185,9 +219,6 @@ class TimeSteppingSystem(Discipline):
 
             if self.cache is not None:
                 self.cache.clear()
-                self.__mda.cache.clear()
-                for mda in self.__mda.inner_mdas:
-                    mda.cache.clear()
 
         if self.__do_step:
             # At initial time,
@@ -203,6 +234,23 @@ class TimeSteppingSystem(Discipline):
         return super().execute(input_data)
 
     def _run(self, input_data: StrKeyMapping) -> StrKeyMapping | None:
+        if self.__time_manager.is_initial and self.__mda_max_iter_at_t0 > 0:
+            max_mda_iter = self.__mda.settings.max_mda_iter
+            for mda in self.__mda.inner_mdas:
+                mda.settings.max_mda_iter = self.__mda_max_iter_at_t0
+
+            for fmu_discipline in self.__fmu_disciplines:
+                fmu_discipline.set_default_execution(initialize_only=True)
+
+            self.__mda.execute()
+            for mda in self.__mda.inner_mdas:
+                mda.settings.max_mda_iter = max_mda_iter
+
+            for fmu_discipline in self.__fmu_disciplines:
+                fmu_discipline.set_default_execution(initialize_only=False)
+
+        input_data = dict(input_data)
+        input_data.update(self.__mda.io.data)
         if self.__do_step:
             self.__simulate_one_time_step(input_data)
             self.io.data.update(self.__mda.io.data)
@@ -224,8 +272,6 @@ class TimeSteppingSystem(Discipline):
             self.__simulate_one_time_step(input_data)
             local_data_history.append(copy(self.__mda.io.data))
             input_data = self.__mda.io.data
-            for inner_mda in self.__mda.inner_mdas:
-                inner_mda.cache.clear()
 
         # The different time steps are concatenated when the values are NumPy arrays.
         # Given a variable,